Move FakeProducerSideService.h to OrbitProducer target
This is so that in can be used in tests by other modules (specifically,
`OrbitVulkanLayer`).
